NextGen-FaceRecognition/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.2
opencv-python>=4.5.0
numpy>=1.21.0
requests>=2.25.0
Pillow>=8.0.0